+++ sdk-extensions/incubator/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider
@@ -1,1 +1,1 @@
-io.opentelemetry.sdk.autoconfigure.internal.ServiceInstanceIdResourceProvider
+io.opentelemetry.sdk.extension.incubator.resources.ServiceInstanceIdResourceProvider
